backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
dj-database-url
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
print("DATABASE_URL:", os.getenv('DATABASE_URL'))